pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finopsguard"
version = "0.1.0"
description = "FinOps Guard agent skeleton"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0",
  "requests>=2.31.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["finopsguard*"]
